# bff-web/src/main/java/com/bff/web/application/api/BffCustomerDeleteControllerApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bff.web.application.api;

import com.bff.web.domain.data.ResponseError;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2021-08-25T10:27:54.842-05:00[America/Bogota]")
@Validated
@Api(value = "BffCustomerDeleteController", description = "the BffCustomerDeleteController API")
public interface BffCustomerDeleteControllerApi {

    /**
     * DELETE /bff/crm/{documentNumber} : Delete customer by documentNumber
     *
     * @param documentNumber document number for customer (required)
     * @return NO_CONTENT (status code 204)
     *         or NOT_FOUND (status code 404)
     */
    @ApiOperation(value = "Delete customer by documentNumber", nickname = "bffDeleteCustomerInformation", notes = "", tags={ "bff-customer-delete-controller", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "NO_CONTENT"),
        @ApiResponse(code = 404, message = "NOT_FOUND", response = ResponseError.class) })
    @ApiImplicitParams({
    })
    @DeleteMapping(
        value = "/bff/crm/{documentNumber}",
        produces = { "application/json" }
    )
    @CrossOrigin(origins = "*")
    ResponseEntity<Void> bffDeleteCustomerInformation(@ApiParam(value = "document number for customer",required=true) @PathVariable("documentNumber") String documentNumber);

}