# bff-web/src/main/java/com/bff/web/application/api/BffRestrictiveListAuthenticateMethodApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bff.web.application.api;

import com.bff.web.domain.data.AuthenticationRequest;
import com.bff.web.domain.data.AuthenticationResponse;
import com.bff.web.domain.data.ResponseError;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;

@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2021-08-25T10:27:54.842-05:00[America/Bogota]")
@Validated
@Api(value = "BffRestrictiveListAuthenticateMethod", description = "the BffRestrictiveListAuthenticateMethod API")
@CrossOrigin(origins = "*")
public interface BffRestrictiveListAuthenticateMethodApi {

    /**
     * POST /bff/restrictive-list/authenticate
     *
     * @param authenticationRequest  (required)
     * @return OK (status code 200)
     *         or UNAUTHORIZED (status code 401)
     */
    @ApiOperation(value = "", nickname = "bffRestrictiveListGenerateAuthentication", notes = "", response = AuthenticationResponse.class, tags={ "bff-restrictive-list-authenticate-method", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = AuthenticationResponse.class),
        @ApiResponse(code = 401, message = "UNAUTHORIZED", response = ResponseError.class) })
    @ApiImplicitParams({
    })
    @PostMapping(
        value = "/bff/restrictive-list/authenticate",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<AuthenticationResponse> bffRestrictiveListGenerateAuthentication(@ApiParam(value = "" ,required=true )  @Valid @RequestBody AuthenticationRequest authenticationRequest);

}