# bff-web/src/main/java/com/bff/web/application/api/BffRestrictiveListControllerApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bff.web.application.api;

import com.bff.web.domain.data.ClientDTO;
import com.bff.web.domain.data.ClientRequest;
import com.bff.web.domain.data.ResponseError;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;

@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2021-08-25T10:27:54.842-05:00[America/Bogota]")
@Validated
@Api(value = "BffRestrictiveListController", description = "the BffRestrictiveListController API")
public interface BffRestrictiveListControllerApi {

    /**
     * POST /bff/restrictive-list : Save customer
     *
     * @param clientRequest  (required)
     * @return OK (status code 200)
     *         or BAD_REQUEST (status code 400)
     *         or UNAUTHORIZED (status code 401)
     */
    @ApiOperation(value = "Save customer", nickname = "bffRestrictiveListValidateClient", notes = "", response = ClientDTO.class, tags={ "bff-restrictive-list-controller", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = ClientDTO.class),
        @ApiResponse(code = 400, message = "BAD_REQUEST", response = ResponseError.class),
        @ApiResponse(code = 401, message = "UNAUTHORIZED", response = Object.class) })
    @ApiImplicitParams({
    @ApiImplicitParam(name = "security", value = "token authorization", required=true, dataType = "String", paramType = "header")
    })
    @PostMapping(
        value = "/bff/restrictive-list",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    @CrossOrigin(origins = "*")
    ResponseEntity<ClientDTO> bffRestrictiveListValidateClient(@ApiParam(value = "" ,required=true )  @Valid @RequestBody ClientRequest clientRequest,@RequestHeader(value="security") String securityAuthorization);

}